SL-480: add patient to queue action
